pre-extractions
- extract Play / Build components within Start
- extract Store context and provider into contexts.js
