+++ src/components/Grid.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useContext } from "react";
 import { observer } from "mobx-react";
 import { TeamMember, TeamMemberPlaceholder } from "./TeamMember";
-import { StoreContext } from "../model/store.js";
+import { StoreContext } from "../model/contexts.js";
 import { positionFor } from "../model/grid/grid";
 import "./Grid.scss";
