+++ src/components/Progress.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useContext } from "react";
 import { observer } from "mobx-react";
-import { StoreContext } from "../model/store.js";
+import { StoreContext } from "../model/contexts.js";
 
 export const Progress = observer(() => {
   const { store } = useContext(StoreContext);
