+++ src/components/TeamsMini.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useContext } from "react";
 import { observer } from "mobx-react";
-import { StoreContext } from "../model/store.js";
+import { StoreContext } from "../model/contexts.js";
 import "./animate_team_selection.scss";
 import { ScoreFaceIcon } from "./Scoring.js";
