+++ src/pages/Start.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useHistory, Link } from "react-router-dom";
 import { randomEasyProblem, randomHardProblem } from "../model/problem";
 
-export function Start() {
+function Play() {
   const history = useHistory();
 
   function visitRandomGameLink(problem) {
@@ -11,54 +11,70 @@ export function Start() {
   }
   return (
     <>
+      <h1 className="title">Play</h1>
+      <div className="buttons">
+        <button
+          className="button is-link is-light"
+          onClick={() => visitRandomGameLink(randomEasyProblem())}
+        >
+          <span className="icon">
+            <i className="fas fa-play"></i>
+          </span>
+          <span>Random Easy Game</span>
+        </button>
+
+        <button
+          className="button is-link is-light"
+          onClick={() => visitRandomGameLink(randomHardProblem())}
+        >
+          <span className="icon">
+            <i className="fas fa-play"></i>
+          </span>
+          <span>Random Hard Game</span>
+        </button>
+      </div>
+    </>
+  );
+}
+
+function Build() {
+  return (
+    <>
+      <h1 className="title">Build</h1>
+      <div className="buttons">
+        <button className="button is-link is-light">
+          <Link to="/build/5x5">
+            <span className="icon">
+              <i className="fas fa-chess-board"></i>
+            </span>
+            <span>5 x 5</span>
+          </Link>
+        </button>
+        <button className="button is-link is-light">
+          <Link to="/build/10x10">
+            <span className="icon">
+              <i className="fas fa-chess-board"></i>
+            </span>
+            <span>10 x 10</span>
+          </Link>
+        </button>
+      </div>
+    </>
+  );
+}
+
+export function Start() {
+  return (
+    <>
       <section className="hero is-medium is-primary is-bold">
         <div className="hero-body">
           <div className="container">
             <div className="columns">
               <div className="column">
-                <h1 className="title">Play</h1>
-                <div className="buttons">
-                  <button
-                    className="button is-link is-light"
-                    onClick={() => visitRandomGameLink(randomEasyProblem())}
-                  >
-                    <span className="icon">
-                      <i className="fas fa-play"></i>
-                    </span>
-                    <span>Random Easy Game</span>
-                  </button>
-
-                  <button
-                    className="button is-link is-light"
-                    onClick={() => visitRandomGameLink(randomHardProblem())}
-                  >
-                    <span className="icon">
-                      <i className="fas fa-play"></i>
-                    </span>
-                    <span>Random Hard Game</span>
-                  </button>
-                </div>
+                <Play />
               </div>
               <div className="column">
-                <h1 className="title">Build</h1>
-                <div className="buttons">
-                  <button className="button is-link is-light">
-                    <Link to="/build/5x5">
-                      <span className="icon">
-                        <i className="fas fa-chess-board"></i>
-                      </span>
-                      <span>5 x 5</span>
-                    </Link>
-                  </button>
-                  <button className="button is-link is-light">
-                    <Link to="/build/10x10">
-                      <span className="icon">
-                        <i className="fas fa-chess-board"></i>
-                      </span>
-                      <span>10 x 10</span>
-                    </Link>
-                  </button>
-                </div>
+                <Build />
               </div>
             </div>
           </div>
